Change type precMm Change constants

+++ src/main/java/com/mtxsoftware/model/forecast/parts/Hour.java
@@ -68,7 +68,7 @@ public class Hour {
     @JsonProperty("soil_moisture")
     private Double soilMoisture;
     @JsonProperty("prec_mm")
-    private Integer precMm;//	Прогнозируемое количество осадков (в мм).
+    private Double precMm;//	Прогнозируемое количество осадков (в мм).
     @JsonProperty("prec_period")
     private Integer precPeriod;//Прогнозируемый период осадков (в минутах).
     @JsonProperty("prec_prob")
@@ -233,12 +233,12 @@ public class Hour {
     }
 
     @JsonProperty("prec_mm")
-    public Integer getPrecMm() {
+    public Double getPrecMm() {
         return precMm;
     }
 
     @JsonProperty("prec_mm")
-    public void setPrecMm(Integer precMm) {
+    public void setPrecMm(Double precMm) {
         this.precMm = precMm;
     }
